requirements.txt
pillow
pygame
mutagen
pydub